+++ NEWS.md
@@ -71,6 +71,10 @@
   `sinpi2k`, `sinuhk_deg`, `sinpi2ur`,
   `cospi2k`, `cosuhk_deg`.
 
+- Support has been added to <stdfix.h> for functions reading fixed-point
+  values from program space like `pgm_read_*` and `pgm_read_*_far`, where
+  `*` stands for the fixed-point constant suffix.
+
 - Support has been added for fixed-point to decimal ASCII conversions:
   `ktoa`, `uktoa`, `hktoa`, `uhktoa`, `rtoa`, `urtoa`, `hrtoa`, `uhrtoa`.